pyproject.toml
[project]
name = "agente-erpnext-g"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index-core>=0.14.19",
    "llama-index-embeddings-ollama>=0.9.0",
    "llama-index-llms-ollama>=0.10.1",
    "llama-index-readers-file>=0.6.0",
    "nicegui>=3.9.0",
    "python-dotenv>=1.2.2",
    "watchdog>=6.0.0",
]
